Move appointment services into single file
